Refactored CRN_Data and CalendarView
Refactored CRN_Data and CalendarView classes to adjust new field days that replaced mon, tu, wed, thu, fri. days of the week are stored inside a hasmpa rather than individial fields.

-gobii vivagananda
