+++ app/src/main/java/com/example/rito/groupapp/ViewRemoveCourseRegistrationActivity.java
@@ -2,7 +2,6 @@ package com.example.rito.groupapp;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.ServiceConnection;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v7.app.AppCompatActivity;
@@ -18,15 +17,15 @@ import android.view.ViewGroup;
 import android.widget.AbsListView;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.rito.groupapp.old.CRN;
+import com.example.rito.groupapp.old.Student;
 import com.firebase.ui.database.FirebaseListAdapter;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
