+++ app/src/main/java/com/example/rito/groupapp/old/Course_Enrollment.java
@@ -1,4 +1,4 @@
-package com.example.rito.groupapp;
+package com.example.rito.groupapp.old;
 
 import com.google.firebase.database.Exclude;
 import com.google.firebase.database.IgnoreExtraProperties;
