+++ app/src/main/java/com/example/rito/groupapp/old/Course_Schedule.java
@@ -1,6 +1,6 @@
 
 
-package com.example.rito.groupapp;
+package com.example.rito.groupapp.old;
 
 import com.google.firebase.database.Exclude;
 import com.google.firebase.database.IgnoreExtraProperties;
