+++ app/src/test/java/com/example/rito/groupapp/Course_ScheduleTest.java
@@ -1,6 +1,6 @@
 package com.example.rito.groupapp;
 
-import com.example.rito.groupapp.*;
+import com.example.rito.groupapp.old.Course_Schedule;
 
 
 import org.junit.After;
