+++ app/src/test/java/com/example/rito/groupapp/StudentTest.java
@@ -1,5 +1,5 @@
 package com.example.rito.groupapp;
-import com.example.rito.groupapp.*;
+import com.example.rito.groupapp.old.Student;
 
 
 import org.junit.After;
